refactor(actions): migrate copy_clipboard + run_command + paste_snippet handlers
Phase 1A action-handler slice 3. Three handlers in one PR — they
share the same shape (each pulls one match arm out of
execute_action) and none has substantial new logic to test.

What lands:
- handlers/copy_clipboard.rs — single-shot delegation to
  ClipboardManager::copy_to_clipboard.
- handlers/run_command.rs — thin delegation to
  actions::system::execute_command.
- handlers/paste_snippet.rs — relocates the entire
  paste_snippet_via_os platform-gated implementation (Win32
  SendInput / Quartz CGEventPost / xdotool) plus the clipboard-
  copy step. lib.rs no longer carries any keystroke-synthesis
  code.
- execute_action's three corresponding arms shrink to one
  delegation each.

Migration progress: 6 of 9 handlers done. Next slice: focus_window
+ focus_browser_tab + reindex_files (the final group).

393 backend tests pass. No new tests; the relocated
paste_snippet_via_os logic was not previously covered by unit
tests (its FFI surface is integration-tested via live-window
captures).
